♻️[Refactor]#76 - 비디오 화면 리팩토링
